+++ src/Exception/HttpException.php
@@ -27,6 +27,7 @@ class HttpException extends Exception
         416 => 'Range Not Satisfiable',
         417 => 'Expectation Failed',
         418 => 'I\'m a teapot',
+        431 => 'Request Header Fields Too Large',
 
         // 5xx
         500 => 'Internal Server Error',