+++ src/Factory/HttpRequestFactory.php
@@ -65,7 +65,7 @@ readonly class HttpRequestFactory
             $headerSize = strlen($header);
 
             if ($headerSize == ($length - 1)) {
-                throw HttpException::createFromCode(400);
+                throw HttpException::createFromCode(431);
             }
 
             $requestSize += $headerSize;
@@ -87,23 +87,93 @@ readonly class HttpRequestFactory
             $headers[$name][] = $value;
         } while (true);
 
-        $contentLength = $headers['Content-Length'][0] ?? 0;
+        $transferEncodings = array_map(
+            fn(string $encoding): string => strtolower(trim($encoding)),
+            explode(',', $headers['Transfer-Encoding'][0] ?? ''),
+        );
 
-        if ($requestSize + $contentLength > $requestSizeMax) {
-            throw HttpException::createFromCode(413);
+        if (in_array('gzip', $transferEncodings)) {
+            throw HttpException::createFromCode(501);
         }
 
-        $content = '';
-        $contentSize = 0;
+        if (in_array('chunked', $transferEncodings)) {
+            if ($protocol !== 'HTTP/1.1') {
+                throw HttpException::createFromCode(501);
+            }
 
-        while ($contentLength != $contentSize) {
-            $chunk = fread($connection, $length);
-            $chunkSize = strlen($chunk);
+            $content = $this->readContentChunked($connection, $requestSizeMax - $requestSize);
+        } else {
+            $contentLength = $headers['Content-Length'][0] ?? 0;
 
-            $content .= $chunk;
-            $contentSize += $chunkSize;
+            if ($requestSize + $contentLength > $requestSizeMax) {
+                throw HttpException::createFromCode(413);
+            }
+
+            $content = $this->readContentByLength($connection, $contentLength);
         }
 
         return $this->create($method, $path, $protocol, $headers, $content);
+    }
+
+    private function readContentByLength($connection, int $size): string
+    {
+        $content = '';
+        $remaining = $size;
+
+        while ($remaining != 0) {
+            $chunk = fread($connection, min($remaining, 8192));
+
+            if ($chunk === false || $chunk === '') {
+                throw new RuntimeException('Failed to read content');
+            }
+
+            $content .= $chunk;
+            $remaining -= strlen($chunk);
+        }
+
+        return $content;
+    }
+
+    /**
+     * @throws HttpException
+     */
+    private function readContentChunked($connection, int $maxSize): string
+    {
+        $content = '';
+        $totalSize = 0;
+
+        while (true) {
+            $chunkSizeLine = fgets($connection, 8192);
+            $chunkSize = hexdec(trim(explode(';', $chunkSizeLine, 2)[0]));
+
+            if ($chunkSize < 0) {
+                throw new RuntimeException('Invalid chunk size');
+            }
+
+            // Финальный чанк - тот, что имеет размер 0 (хотя на самом деле там дальше trailer headers)
+            if ($chunkSize === 0) {
+                do {
+                    // Trailer - опциональные HTTP заголовки после финального чанка
+                    $trailer = fgets($connection, 8192);
+                } while (!in_array($trailer, ["\r\n", "\n"]));
+
+                break;
+            }
+
+            if ($totalSize + $chunkSize > $maxSize) {
+                throw HttpException::createFromCode(413);
+            }
+
+            $content .= $this->readContentByLength($connection, $chunkSize);
+
+            $totalSize += $chunkSize;
+
+            $crlf = fread($connection, 2);
+            if ($crlf != "\r\n") {
+                throw new RuntimeException('Missing CRLF after chunk data');
+            }
+        }
+
+        return $content;
     }
 }