+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning v2.0.0](http://semver.org/).
 ## [0.3.0] - 2026-01-30
 ### Added
 - Welcome page
+- Processing header "Transfer-Encoding: chunked"
 - Global configuration using XML file
 - Prefork server model
 - Signal processing